* MyFork/master:
  Correct type using for 32 and 64 bit architecture

+++ CustomIOS7AlertView/CustomIOS7AlertView/View/CustomIOS7AlertView.h
@@ -28,7 +28,7 @@
 @property (nonatomic, retain) NSArray *buttonTitles;
 @property (nonatomic, assign) BOOL useMotionEffects;
 
-@property (copy) void (^onButtonTouchUpInside)(CustomIOS7AlertView *alertView, int buttonIndex) ;
+@property (copy) void (^onButtonTouchUpInside)(CustomIOS7AlertView *alertView, NSInteger buttonIndex) ;
 
 - (id)init;
 
@@ -41,7 +41,7 @@
 - (void)close;
 
 - (IBAction)customIOS7dialogButtonTouchUpInside:(id)sender;
-- (void)setOnButtonTouchUpInside:(void (^)(CustomIOS7AlertView *alertView, int buttonIndex))onButtonTouchUpInside;
+- (void)setOnButtonTouchUpInside:(void (^)(CustomIOS7AlertView *alertView, NSInteger buttonIndex))onButtonTouchUpInside;
 
 - (void)deviceOrientationDidChange: (NSNotification *)notification;
 - (void)dealloc;
